Break apart selection.ts into parse and assemble modules.
