+++ src/compile/facet.ts
@@ -21,7 +21,7 @@ import {parseChildrenLayoutSize} from './layoutsize/parse';
 import {Model, ModelWithField} from './model';
 import {RepeaterValue, replaceRepeaterInFacet} from './repeater';
 import {assembleDomain, getFieldFromDomain} from './scale/domain';
-import {assembleFacetSignals} from './selection/selection';
+import {assembleFacetSignals} from './selection/assemble';
 
 export function facetSortFieldName(
   fieldDef: FacetFieldDef<string>,
