+++ test/compile/selection/facets.test.ts
@@ -1,7 +1,8 @@
 /* tslint:disable quotemark */
 
 import {FacetModel} from '../../../src/compile/facet';
-import * as selection from '../../../src/compile/selection/selection';
+import {unitName} from '../../../src/compile/selection';
+import {assembleFacetSignals} from '../../../src/compile/selection/assemble';
 import {UnitModel} from '../../../src/compile/unit';
 import {parseModel} from '../../util';
 
@@ -37,7 +38,7 @@ describe('Faceted Selections', () => {
   const unit = model.children[0].children[1] as UnitModel;
 
   it('should assemble a facet signal', () => {
-    expect(selection.assembleFacetSignals(model as FacetModel, [])).toContainEqual({
+    expect(assembleFacetSignals(model as FacetModel, [])).toContainEqual({
       name: 'facet',
       value: {},
       on: [
@@ -50,7 +51,7 @@ describe('Faceted Selections', () => {
   });
 
   it('should name the unit with the facet keys', () => {
-    expect(selection.unitName(unit)).toEqual(
+    expect(unitName(unit)).toEqual(
       `"child_layer_1" + '__facet_row_' + (facet["bin_maxbins_6_X"]) + '__facet_column_' + (facet["Series"])`
     );
   });
