+++ test/compile/selection/multi.test.ts
@@ -1,6 +1,11 @@
 /* tslint:disable quotemark */
+import {
+  assembleUnitSelectionData,
+  assembleUnitSelectionMarks,
+  assembleUnitSelectionSignals
+} from '../../../src/compile/selection/assemble';
 import multi from '../../../src/compile/selection/multi';
-import * as selection from '../../../src/compile/selection/selection';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import {parseUnitModelWithScale} from '../../util';
 
 describe('Multi Selection', () => {
@@ -13,7 +18,7 @@ describe('Multi Selection', () => {
     }
   });
 
-  const selCmpts = (model.component.selection = selection.parseUnitSelection(model, {
+  const selCmpts = (model.component.selection = parseUnitSelection(model, {
     one: {type: 'multi'},
     two: {
       type: 'multi',
@@ -137,13 +142,13 @@ describe('Multi Selection', () => {
       }
     ]);
 
-    const signals = selection.assembleUnitSelectionSignals(model, []);
+    const signals = assembleUnitSelectionSignals(model, []);
     expect(signals).toEqual(expect.arrayContaining([...oneSg, ...twoSg, ...threeSg, ...fourSg, ...fiveSg]));
   });
 
   it('builds unit datasets', () => {
     const data: any[] = [];
-    expect(selection.assembleUnitSelectionData(model, data)).toEqual([
+    expect(assembleUnitSelectionData(model, data)).toEqual([
       {name: 'one_store'},
       {name: 'two_store'},
       {name: 'thr_ee_store'},
@@ -155,6 +160,6 @@ describe('Multi Selection', () => {
   it('leaves marks alone', () => {
     const marks: any[] = [];
     model.component.selection = {one: selCmpts['one']};
-    expect(selection.assembleUnitSelectionMarks(model, marks)).toEqual(marks);
+    expect(assembleUnitSelectionMarks(model, marks)).toEqual(marks);
   });
 });
