+++ test/compile/selection/nearest.test.ts
@@ -1,6 +1,6 @@
 /* tslint:disable quotemark */
 
-import * as selection from '../../../src/compile/selection/selection';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import nearest from '../../../src/compile/selection/transforms/nearest';
 import * as log from '../../../src/log';
 import {duplicate} from '../../../src/util';
@@ -17,7 +17,7 @@ function getModel(markType: any) {
   });
   model.parseScale();
   model.parseMarkGroup();
-  model.component.selection = selection.parseUnitSelection(model, {
+  model.component.selection = parseUnitSelection(model, {
     one: {type: 'single', nearest: true},
     two: {type: 'multi', nearest: true},
     three: {type: 'interval'},
