+++ test/compile/selection/predicate.test.ts
@@ -2,10 +2,9 @@
 
 import {nonPosition} from '../../../src/compile/mark/mixins';
 import {expression} from '../../../src/compile/predicate';
-import * as selection from '../../../src/compile/selection/selection';
+import {assembleSelectionPredicate as predicate} from '../../../src/compile/selection/assemble';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import {parseUnitModel} from '../../util';
-
-const predicate = selection.selectionPredicate;
 
 describe('Selection Predicate', () => {
   const model = parseUnitModel({
@@ -34,7 +33,7 @@ describe('Selection Predicate', () => {
 
   model.parseScale();
 
-  model.component.selection = selection.parseUnitSelection(model, {
+  model.component.selection = parseUnitSelection(model, {
     one: {type: 'single'},
     two: {type: 'multi', resolve: 'union'},
     'thr-ee': {type: 'interval', resolve: 'intersect'},
