+++ test/compile/selection/single.test.ts
@@ -1,5 +1,11 @@
 /* tslint:disable quotemark */
-import * as selection from '../../../src/compile/selection/selection';
+import {
+  assembleTopLevelSignals,
+  assembleUnitSelectionData,
+  assembleUnitSelectionMarks,
+  assembleUnitSelectionSignals
+} from '../../../src/compile/selection/assemble';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import single from '../../../src/compile/selection/single';
 import {parseUnitModelWithScale} from '../../util';
 
@@ -15,7 +21,7 @@ describe('Single Selection', () => {
 
   model.parseScale();
 
-  const selCmpts = (model.component.selection = selection.parseUnitSelection(model, {
+  const selCmpts = (model.component.selection = parseUnitSelection(model, {
     one: {type: 'single'},
     two: {
       type: 'single',
@@ -105,7 +111,7 @@ describe('Single Selection', () => {
       }
     ]);
 
-    const signals = selection.assembleUnitSelectionSignals(model, []);
+    const signals = assembleUnitSelectionSignals(model, []);
     expect(signals).toEqual(expect.arrayContaining([...oneSg, ...twoSg, ...threeSg, ...fourSg]));
   });
 
@@ -116,7 +122,7 @@ describe('Single Selection', () => {
     const twoExpr = single.modifyExpr(model, selCmpts['two']);
     expect(twoExpr).toEqual('two_tuple, {unit: ""}');
 
-    const signals = selection.assembleUnitSelectionSignals(model, []);
+    const signals = assembleUnitSelectionSignals(model, []);
     expect(signals).toEqual(
       expect.arrayContaining([
         {
@@ -132,7 +138,7 @@ describe('Single Selection', () => {
   });
 
   it('builds top-level signals', () => {
-    const signals = selection.assembleTopLevelSignals(model, []);
+    const signals = assembleTopLevelSignals(model, []);
     expect(signals).toEqual(
       expect.arrayContaining([
         {
@@ -154,7 +160,7 @@ describe('Single Selection', () => {
 
   it('builds unit datasets', () => {
     const data: any[] = [];
-    expect(selection.assembleUnitSelectionData(model, data)).toEqual([
+    expect(assembleUnitSelectionData(model, data)).toEqual([
       {name: 'one_store'},
       {name: 'two_store'},
       {name: 'thr_ee_store'},
@@ -165,6 +171,6 @@ describe('Single Selection', () => {
   it('leaves marks alone', () => {
     const marks: any[] = [];
     model.component.selection = {one: selCmpts['one']};
-    expect(selection.assembleUnitSelectionMarks(model, marks)).toEqual(marks);
+    expect(assembleUnitSelectionMarks(model, marks)).toEqual(marks);
   });
 });
