+++ test/compile/selection/toggle.test.ts
@@ -1,6 +1,7 @@
 /* tslint:disable quotemark */
 
-import * as selection from '../../../src/compile/selection/selection';
+import {assembleUnitSelectionSignals} from '../../../src/compile/selection/assemble';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import toggle from '../../../src/compile/selection/transforms/toggle';
 import {parseUnitModel} from '../../util';
 
@@ -15,7 +16,7 @@ describe('Toggle Selection Transform', () => {
   });
 
   model.parseScale();
-  const selCmpts = (model.component.selection = selection.parseUnitSelection(model, {
+  const selCmpts = (model.component.selection = parseUnitSelection(model, {
     one: {type: 'multi'},
     two: {
       type: 'multi',
@@ -68,7 +69,7 @@ describe('Toggle Selection Transform', () => {
       }
     ]);
 
-    const signals = selection.assembleUnitSelectionSignals(model, []);
+    const signals = assembleUnitSelectionSignals(model, []);
     expect(signals).toEqual(expect.arrayContaining([...oneSg, ...twoSg]));
   });
 
@@ -81,7 +82,7 @@ describe('Toggle Selection Transform', () => {
       'two_toggle ? null : two_tuple, two_toggle ? null : {unit: ""}, two_toggle ? two_tuple : null'
     );
 
-    const signals = selection.assembleUnitSelectionSignals(model, []);
+    const signals = assembleUnitSelectionSignals(model, []);
     expect(signals).toEqual(
       expect.arrayContaining([
         {
