+++ test/compile/selection/translate.test.ts
@@ -1,7 +1,8 @@
 /* tslint:disable quotemark */
 
 import {selector as parseSelector} from 'vega-event-selector';
-import * as selection from '../../../src/compile/selection/selection';
+import {assembleUnitSelectionSignals} from '../../../src/compile/selection/assemble';
+import {parseUnitSelection} from '../../../src/compile/selection/parse';
 import translate from '../../../src/compile/selection/transforms/translate';
 import {ScaleType} from '../../../src/scale';
 import {parseUnitModel} from '../../util';
@@ -17,7 +18,7 @@ function getModel(xscale?: ScaleType, yscale?: ScaleType) {
   });
 
   model.parseScale();
-  const selCmpts = selection.parseUnitSelection(model, {
+  const selCmpts = parseUnitSelection(model, {
     one: {
       type: 'single'
     },
@@ -65,7 +66,7 @@ describe('Translate Selection Transform', () => {
 
     it('builds them for default invocation', () => {
       model.component.selection = {four: selCmpts['four']};
-      const signals = selection.assembleUnitSelectionSignals(model, []);
+      const signals = assembleUnitSelectionSignals(model, []);
       expect(signals).toEqual(
         expect.arrayContaining([
           {
@@ -94,7 +95,7 @@ describe('Translate Selection Transform', () => {
 
     it('builds them for custom events', () => {
       model.component.selection = {five: selCmpts['five']};
-      const signals = selection.assembleUnitSelectionSignals(model, []);
+      const signals = assembleUnitSelectionSignals(model, []);
       expect(signals).toEqual(
         expect.arrayContaining([
           {
@@ -126,7 +127,7 @@ describe('Translate Selection Transform', () => {
 
     it('builds them for scale-bound intervals', () => {
       model.component.selection = {six: selCmpts['six']};
-      const signals = selection.assembleUnitSelectionSignals(model, []);
+      const signals = assembleUnitSelectionSignals(model, []);
       expect(signals).toEqual(
         expect.arrayContaining([
           {
@@ -158,7 +159,7 @@ describe('Translate Selection Transform', () => {
     it('always builds panLinear exprs for brushes', () => {
       const {model, selCmpts} = getModel();
       model.component.selection = {four: selCmpts['four']};
-      let signals = selection.assembleUnitSelectionSignals(model, []);
+      let signals = assembleUnitSelectionSignals(model, []);
       expect(signals.filter(s => s.name === 'four_x')[0].on).toContainEqual({
         events: {signal: 'four_translate_delta'},
         update:
@@ -173,7 +174,7 @@ describe('Translate Selection Transform', () => {
 
       const model2 = getModel('log', 'pow').model;
       model2.component.selection = {four: selCmpts['four']};
-      signals = selection.assembleUnitSelectionSignals(model2, []);
+      signals = assembleUnitSelectionSignals(model2, []);
       expect(signals.filter(s => s.name === 'four_x')[0].on).toContainEqual({
         events: {signal: 'four_translate_delta'},
         update:
@@ -190,7 +191,7 @@ describe('Translate Selection Transform', () => {
     it('builds panLinear exprs for scale-bound intervals', () => {
       const {model, selCmpts} = getModel();
       model.component.selection = {six: selCmpts['six']};
-      const signals = selection.assembleUnitSelectionSignals(model, []);
+      const signals = assembleUnitSelectionSignals(model, []);
 
       expect(signals.filter(s => s.name === 'six_Horsepower')[0].on).toContainEqual({
         events: {signal: 'six_translate_delta'},
@@ -206,7 +207,7 @@ describe('Translate Selection Transform', () => {
     it('builds panLog/panPow exprs for scale-bound intervals', () => {
       const {model, selCmpts} = getModel('log', 'pow');
       model.component.selection = {six: selCmpts['six']};
-      const signals = selection.assembleUnitSelectionSignals(model, []);
+      const signals = assembleUnitSelectionSignals(model, []);
 
       expect(signals.filter(s => s.name === 'six_Horsepower')[0].on).toContainEqual({
         events: {signal: 'six_translate_delta'},
